[TS-25] Refactor: Chatting폴더안에 각 모달별로 폴더만들어 넣음
